Pages

+++ src/TossErp.Shared/DTOs/FinancialReportItemDto.cs
@@ -4,6 +4,6 @@ public class FinancialReportItemDto
 {
     public string Category { get; set; } = string.Empty;
     public decimal Amount { get; set; }
-    public decimal Percentage { get; set; }
+    public double Percentage { get; set; }
     public string Type { get; set; } = string.Empty;
 }
